+++ Source/Objects/Hardware/USB/LakeShore336/ORLakeShore336Model.m
@@ -597,7 +597,7 @@ NSString* ORLakeShore336PollTimeChanged         = @"ORLakeShore336PollTimeChange
     }
 }
 
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount
 {
     if(inNetSocket == socket){
 		NSString* theString = [inNetSocket readString:NSASCIIStringEncoding];
